+++ CHANGELOG.md
@@ -5,9 +5,9 @@
 - Added a new service layer to opearate git commands.
 - Added a new service layer to talk with GitLab API.
 - Added new methods to get info from Git and GitLab.
+- Added Personal Access Token flow providing menu options to save and delete GitLab PAT.
 - Implemented MR link on status bar.
 - Implemented pipeline status on status bar.
-- Added Personal Access Token flow providing menu options to save and delete GitLab PAT.
 
 ### Changed
 - Removed `gitlab.userId` necessity.